Added templates for croppie
