+++ contentcuration/contentcuration/static/js/edit_channel/file_upload/hbtemplates/file_upload_format_item.handlebars
@@ -1,6 +1,6 @@
 <div class="row uploaded_format_item format_item">
-    <div class="inline col-xs-2 preview_section text-center"><div class="preview_thumbnail"></div></div>
-    <div class="inline col-xs-9 format_section format_metadata_section">
+    <div class="inline col-xs-3 preview_section text-center"><div class="preview_thumbnail"></div></div>
+    <div class="inline col-xs-8 format_section format_metadata_section">
         <input type="text" placeholder="Title *" class="name_content_input" value="{{node.title}}"/>
         <h4 class="format_editor_header">
             Files
