+++ contentcuration/contentcuration/static/js/edit_channel/file_upload/hbtemplates/file_upload_inline_item.handlebars
@@ -1,6 +1,6 @@
 <div class="row">
-    <div class="inline col-xs-3 preview_section"><div class="preview_thumbnail"></div></div>
-    <div class="inline col-xs-9 format_section format_metadata_section">
+    <div class="inline col-xs-4 preview_section"><div class="preview_thumbnail"></div></div>
+    <div class="inline col-xs-8 format_section format_metadata_section">
     {{#if has_files}}
         <h4 class="format_editor_header">
             Files
