+++ contentcuration/contentcuration/static/js/edit_channel/image/hbtemplates/image_upload_modal.handlebars
