+++ contentcuration/contentcuration/static/js/edit_channel/image/hbtemplates/thumbnail_dropzone.handlebars
