+++ contentcuration/contentcuration/static/js/edit_channel/image/hbtemplates/thumbnail_generator_preview.handlebars
