+++ contentcuration/contentcuration/static/js/edit_channel/image/hbtemplates/thumbnail_preview.handlebars
