+++ contentcuration/contentcuration/static/js/edit_channel/image/hbtemplates/thumbnail_upload.handlebars
@@ -7,6 +7,7 @@
             {{else}}&nbsp;
             {{/if}}
             <span class="image_option glyphicon glyphicon-remove remove_image" aria-hidden="true"></span>
+            <span class="image_option glyphicon glyphicon-ok get_image" aria-hidden="true"></span>
         </div>
     </div>
 </div>
