+++ contentcuration/contentcuration/static/less/styles.less
@@ -253,3 +253,5 @@ body {
 	background:#e9e9e9;
 	font-weight:bold;
 }
+
+.crop-img .cr-viewport{ border: none !important; }
